auth-service/internal/utils/paseto_test.go
package utils_test

import (
	"kleimak/ainomy-authentication-service/internal/utils"
	"testing"
	"time"

	"github.com/stretchr/testify/require"
)

func TestPasetoMaker(t *testing.T) {
	maker, err := utils.NewPasetoMaker(utils.RandomString(32))
	require.NoError(t, err)

	address, err := utils.RandomEthereumAddress()
	require.NoError(t, err)

	duration := time.Minute
	issuedAt := time.Now()
	expiredAt := issuedAt.Add(duration)
	token, err := maker.CreateToken(address, duration)

	require.NoError(t, err)
	require.NotEmpty(t, token)

	payload, err := maker.VerifyToken(token)
	require.NoError(t, err)
	require.NotEmpty(t, token)

	require.Equal(t, address, payload.Address)
	require.WithinDuration(t, issuedAt, payload.IssuedAt, time.Second)
	require.WithinDuration(t, expiredAt, payload.ExpiredAt, time.Second)
}
